+++ src/main/java/roomescape/repository/JdbcReservationRepository.java
@@ -104,7 +104,6 @@ public class JdbcReservationRepository implements ReservationRepository {
     @Override
     public Boolean existId(Long id) {
         String sql = "SELECT EXISTS (SELECT 1 FROM reservation WHERE id = ?)";
-        ;
         return jdbcTemplate.queryForObject(sql, Boolean.class, id);
     }
